requirement.txt
tensorflow==2.15.0
keras==2.15.0
numpy
matplotlib
pillow
opencv-python
scikit-learn
streamlit
ultralytics